pocs/poc/requirements.txt
accelerate==1.12.0
anyio==4.12.1
bitsandbytes==0.49.1
certifi==2026.1.4
click==8.3.1
contourpy==1.3.3
cycler==0.12.1
filelock==3.20.0
fonttools==4.61.1
fsspec==2025.12.0
h11==0.16.0
hf-xet==1.2.0
httpcore==1.0.9
httpx==0.28.1
huggingface-hub==1.3.4
idna==3.11
jinja2==3.1.6
joblib==1.5.3
kiwisolver==1.4.9
markupsafe==2.1.5
matplotlib==3.10.8
mpmath==1.3.0
networkx==3.6.1
numpy==2.3.5
nvidia-cublas-cu12==12.1.3.1
nvidia-cuda-cupti-cu12==12.1.105
nvidia-cuda-nvrtc-cu12==12.1.105
nvidia-cuda-runtime-cu12==12.1.105
nvidia-cudnn-cu12==9.1.0.70
nvidia-cufft-cu12==11.0.2.54
nvidia-curand-cu12==10.3.2.106
nvidia-cusolver-cu12==11.4.5.107
nvidia-cusparse-cu12==12.1.0.106
nvidia-nccl-cu12==2.21.5
nvidia-nvjitlink-cu12==12.9.86
nvidia-nvtx-cu12==12.1.105
packaging==26.0
pandas==3.0.0
pgvector==0.4.2
pillow==12.0.0
psutil==7.2.1
psycopg2-binary==2.9.11
pyparsing==3.3.2
python-dateutil==2.9.0.post0
pyyaml==6.0.3
regex==2026.1.15
safetensors==0.7.0
scikit-learn==1.8.0
scipy==1.17.0
sentence-transformers==5.2.1
shellingham==1.5.4
six==1.17.0
sympy==1.13.1
threadpoolctl==3.6.0
tokenizers==0.22.2
torch==2.5.1+cu121
torchvision==0.20.1+cu121
tqdm==4.67.1
transformers==5.0.0
triton==3.1.0
typer-slim==0.21.1
typing-extensions==4.15.0
